+++ vowpalwabbit/global_data.h
@@ -145,9 +145,9 @@ struct vw {
   bool sequence;
   bool searn;
 
-  size_t base_learner_nb_w; //this stores the current number of "weight vector" required by the based learner, which is used to compute offsets when composing reductions
+  uint32_t base_learner_nb_w; //this stores the current number of "weight vector" required by the based learner, which is used to compute offsets when composing reductions
 
-  size_t stride;
+  uint32_t stride;
 
   std::string per_feature_regularizer_input;
   std::string per_feature_regularizer_output;
@@ -203,7 +203,7 @@ struct vw {
   size_t rank;
 
   //Prediction output
-  v_array<size_t> final_prediction_sink; // set to send global predictions to.
+  v_array<int> final_prediction_sink; // set to send global predictions to.
   int raw_prediction; // file descriptors for text output.
   size_t unique_id; //unique id for each node in the network, id == 0 means extra io.
   size_t total; //total number of nodes
